refactor: move code files to src subdir
Reshuffling the project structure to keep code separate from tests.
