+++ src/streams.js
